Enable IMA and EVM Feature
IMA, EVM Distro feature support added.

Distro feature native will append the IMA support
based on selected machine and it will allows
to generate pre-defined UUIDs.
For IMA, we need predefined UUIDs so that it can
be compared with UUID defined in kernel driver
to identify partition/image on target.

Change-Id: I1bc15086f8959a9dd8529db622f06e9310375a1c
Signed-off-by: quic_venper <quic_venper@quicinc.com>